[refactor] RegisterVC와 RegisterView를 분리
