+++ Soulmate/Soulmate/Presentation/Interface/AuthUseCase.swift
@@ -1,0 +1,12 @@
+//
+//  AuthUseCase.swift
+//  Soulmate
+//
+//  Created by Sangmin Lee on 2022/12/19.
+//
+
+import Foundation
+
+protocol AuthUseCase {
+    func userUid() -> String?
+}
